Share common readFrom/writeTo code in AcknowledgeResponse
