+++ server/src/main/java/org/elasticsearch/action/admin/indices/close/CloseIndexResponse.java
@@ -20,12 +20,8 @@
 package org.elasticsearch.action.admin.indices.close;
 
 import org.elasticsearch.action.support.master.AcknowledgedResponse;
-import org.elasticsearch.common.io.stream.StreamInput;
-import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.xcontent.ConstructingObjectParser;
 import org.elasticsearch.common.xcontent.XContentParser;
-
-import java.io.IOException;
 
 /**
  * A response for a close index action.
@@ -43,18 +39,6 @@ public class CloseIndexResponse extends AcknowledgedResponse {
 
     CloseIndexResponse(boolean acknowledged) {
         super(acknowledged);
-    }
-
-    @Override
-    public void readFrom(StreamInput in) throws IOException {
-        super.readFrom(in);
-        readAcknowledged(in);
-    }
-
-    @Override
-    public void writeTo(StreamOutput out) throws IOException {
-        super.writeTo(out);
-        writeAcknowledged(out);
     }
 
     public static CloseIndexResponse fromXContent(XContentParser parser) {
