+++ server/src/main/java/org/elasticsearch/action/admin/indices/settings/put/UpdateSettingsResponse.java
@@ -20,12 +20,8 @@
 package org.elasticsearch.action.admin.indices.settings.put;
 
 import org.elasticsearch.action.support.master.AcknowledgedResponse;
-import org.elasticsearch.common.io.stream.StreamInput;
-import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.xcontent.ConstructingObjectParser;
 import org.elasticsearch.common.xcontent.XContentParser;
-
-import java.io.IOException;
 
 /**
  * A response for an update index settings action
@@ -44,18 +40,6 @@ public class UpdateSettingsResponse extends AcknowledgedResponse {
 
     UpdateSettingsResponse(boolean acknowledged) {
         super(acknowledged);
-    }
-
-    @Override
-    public void readFrom(StreamInput in) throws IOException {
-        super.readFrom(in);
-        readAcknowledged(in);
-    }
-
-    @Override
-    public void writeTo(StreamOutput out) throws IOException {
-        super.writeTo(out);
-        writeAcknowledged(out);
     }
 
     public static UpdateSettingsResponse fromXContent(XContentParser parser) {
