+++ server/src/main/java/org/elasticsearch/action/ingest/PutPipelineResponse.java
@@ -20,13 +20,9 @@
 package org.elasticsearch.action.ingest;
 
 import org.elasticsearch.action.support.master.AcknowledgedResponse;
-import org.elasticsearch.common.io.stream.StreamInput;
-import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.xcontent.ConstructingObjectParser;
 import org.elasticsearch.common.xcontent.ToXContentObject;
 import org.elasticsearch.common.xcontent.XContentParser;
-
-import java.io.IOException;
 
 public class PutPipelineResponse extends AcknowledgedResponse implements ToXContentObject {
 
@@ -42,18 +38,6 @@ public class PutPipelineResponse extends AcknowledgedResponse implements ToXCont
 
     public PutPipelineResponse(boolean acknowledged) {
         super(acknowledged);
-    }
-
-    @Override
-    public void readFrom(StreamInput in) throws IOException {
-        super.readFrom(in);
-        readAcknowledged(in);
-    }
-
-    @Override
-    public void writeTo(StreamOutput out) throws IOException {
-        super.writeTo(out);
-        writeAcknowledged(out);
     }
 
     public static PutPipelineResponse fromXContent(XContentParser parser) {
