+++ server/src/main/java/org/elasticsearch/action/ingest/WritePipelineResponse.java
@@ -20,10 +20,6 @@
 package org.elasticsearch.action.ingest;
 
 import org.elasticsearch.action.support.master.AcknowledgedResponse;
-import org.elasticsearch.common.io.stream.StreamInput;
-import org.elasticsearch.common.io.stream.StreamOutput;
-
-import java.io.IOException;
 
 public class WritePipelineResponse extends AcknowledgedResponse {
 
@@ -33,17 +29,5 @@ public class WritePipelineResponse extends AcknowledgedResponse {
 
     public WritePipelineResponse(boolean acknowledged) {
         super(acknowledged);
-    }
-
-    @Override
-    public void readFrom(StreamInput in) throws IOException {
-        super.readFrom(in);
-        readAcknowledged(in);
-    }
-
-    @Override
-    public void writeTo(StreamOutput out) throws IOException {
-        super.writeTo(out);
-        writeAcknowledged(out);
     }
 }
