+++ x-pack/plugin/core/src/main/java/org/elasticsearch/license/DeleteLicenseResponse.java
@@ -6,10 +6,6 @@
 package org.elasticsearch.license;
 
 import org.elasticsearch.action.support.master.AcknowledgedResponse;
-import org.elasticsearch.common.io.stream.StreamInput;
-import org.elasticsearch.common.io.stream.StreamOutput;
-
-import java.io.IOException;
 
 public class DeleteLicenseResponse extends AcknowledgedResponse {
 
@@ -19,17 +15,4 @@ public class DeleteLicenseResponse extends AcknowledgedResponse {
     DeleteLicenseResponse(boolean acknowledged) {
         super(acknowledged);
     }
-
-    @Override
-    public void readFrom(StreamInput in) throws IOException {
-        super.readFrom(in);
-        readAcknowledged(in);
-    }
-
-    @Override
-    public void writeTo(StreamOutput out) throws IOException {
-        super.writeTo(out);
-        writeAcknowledged(out);
-    }
-
 }
